+++ src/etl/LoadBalancer.cpp
@@ -21,7 +21,6 @@
 
 #include "data/BackendInterface.hpp"
 #include "etl/ETLHelpers.hpp"
-#include "etl/ETLService.hpp"
 #include "etl/ETLState.hpp"
 #include "etl/Source.hpp"
 #include "feed/SubscriptionManagerInterface.hpp"
@@ -59,11 +58,12 @@ LoadBalancer::make_LoadBalancer(
     boost::asio::io_context& ioc,
     std::shared_ptr<BackendInterface> backend,
     std::shared_ptr<feed::SubscriptionManagerInterface> subscriptions,
-    std::shared_ptr<NetworkValidatedLedgersInterface> validatedLedgers
+    std::shared_ptr<NetworkValidatedLedgersInterface> validatedLedgers,
+    SourceFactory sourceFactory
 )
 {
     return std::make_shared<LoadBalancer>(
-        config, ioc, std::move(backend), std::move(subscriptions), std::move(validatedLedgers)
+        config, ioc, std::move(backend), std::move(subscriptions), std::move(validatedLedgers), std::move(sourceFactory)
     );
 }
 
@@ -72,7 +72,8 @@ LoadBalancer::LoadBalancer(
     boost::asio::io_context& ioc,
     std::shared_ptr<BackendInterface> backend,
     std::shared_ptr<feed::SubscriptionManagerInterface> subscriptions,
-    std::shared_ptr<NetworkValidatedLedgersInterface> validatedLedgers
+    std::shared_ptr<NetworkValidatedLedgersInterface> validatedLedgers,
+    SourceFactory sourceFactory
 )
 {
     auto const forwardingCacheTimeout = config.valueOr<float>("forwarding_cache_timeout", 0.f);
@@ -101,7 +102,7 @@ LoadBalancer::LoadBalancer(
     };
 
     for (auto const& entry : config.array("etl_sources")) {
-        auto source = make_Source(
+        auto source = sourceFactory(
             entry,
             ioc,
             backend,
@@ -116,12 +117,12 @@ LoadBalancer::LoadBalancer(
         );
 
         // checking etl node validity
-        auto const stateOpt = ETLState::fetchETLStateFromSource(source);
+        auto const stateOpt = ETLState::fetchETLStateFromSource(*source);
 
         if (!stateOpt) {
             checkOnETLFailure(fmt::format(
                 "Failed to fetch ETL state from source = {} Please check the configuration and network",
-                source.toString()
+                source->toString()
             ));
         } else if (etlState_ && etlState_->networkID && stateOpt->networkID && etlState_->networkID != stateOpt->networkID) {
             checkOnETLFailure(fmt::format(
@@ -134,7 +135,7 @@ LoadBalancer::LoadBalancer(
         }
 
         sources_.push_back(std::move(source));
-        LOG(log_.info()) << "Added etl source - " << sources_.back().toString();
+        LOG(log_.info()) << "Added etl source - " << sources_.back()->toString();
     }
 
     if (sources_.empty())
@@ -142,8 +143,8 @@ LoadBalancer::LoadBalancer(
 
     // This is made separate from source creation to prevent UB in case one of the sources will call
     // chooseForwardingSource while we are still filling the sources_ vector
-    for (auto& source : sources_) {
-        source.run();
+    for (auto const& source : sources_) {
+        source->run();
     }
 }
 
@@ -158,11 +159,11 @@ LoadBalancer::loadInitialLedger(uint32_t sequence, bool cacheOnly)
     std::vector<std::string> response;
     auto const success = execute(
         [this, &response, &sequence, cacheOnly](auto& source) {
-            auto [data, res] = source.loadInitialLedger(sequence, downloadRanges_, cacheOnly);
+            auto [data, res] = source->loadInitialLedger(sequence, downloadRanges_, cacheOnly);
 
             if (!res) {
                 LOG(log_.error()) << "Failed to download initial ledger."
-                                  << " Sequence = " << sequence << " source = " << source.toString();
+                                  << " Sequence = " << sequence << " source = " << source->toString();
             } else {
                 response = std::move(data);
             }
@@ -180,17 +181,17 @@ LoadBalancer::fetchLedger(uint32_t ledgerSequence, bool getObjects, bool getObje
     GetLedgerResponseType response;
     bool const success = execute(
         [&response, ledgerSequence, getObjects, getObjectNeighbors, log = log_](auto& source) {
-            auto [status, data] = source.fetchLedger(ledgerSequence, getObjects, getObjectNeighbors);
+            auto [status, data] = source->fetchLedger(ledgerSequence, getObjects, getObjectNeighbors);
             response = std::move(data);
             if (status.ok() && response.validated()) {
                 LOG(log.info()) << "Successfully fetched ledger = " << ledgerSequence
-                                << " from source = " << source.toString();
+                                << " from source = " << source->toString();
                 return true;
             }
 
             LOG(log.warn()) << "Could not fetch ledger " << ledgerSequence << ", Reply: " << response.DebugString()
                             << ", error_code: " << status.error_code() << ", error_msg: " << status.error_message()
-                            << ", source = " << source.toString();
+                            << ", source = " << source->toString();
             return false;
         },
         ledgerSequence
@@ -222,7 +223,7 @@ LoadBalancer::forwardToRippled(
 
     std::optional<boost::json::object> response;
     while (numAttempts < sources_.size()) {
-        if (auto res = sources_[sourceIdx].forwardToRippled(request, clientIp, yield)) {
+        if (auto res = sources_[sourceIdx]->forwardToRippled(request, clientIp, yield)) {
             response = std::move(res);
             break;
         }
@@ -242,7 +243,7 @@ LoadBalancer::toJson() const
 {
     boost::json::array ret;
     for (auto& src : sources_)
-        ret.push_back(src.toJson());
+        ret.push_back(src->toJson());
 
     return ret;
 }
@@ -261,23 +262,23 @@ LoadBalancer::execute(Func f, uint32_t ledgerSequence)
         auto& source = sources_[sourceIdx];
 
         LOG(log_.debug()) << "Attempting to execute func. ledger sequence = " << ledgerSequence
-                          << " - source = " << source.toString();
+                          << " - source = " << source->toString();
         // Originally, it was (source->hasLedger(ledgerSequence) || true)
         /* Sometimes rippled has ledger but doesn't actually know. However,
         but this does NOT happen in the normal case and is safe to remove
         This || true is only needed when loading full history standalone */
-        if (source.hasLedger(ledgerSequence)) {
+        if (source->hasLedger(ledgerSequence)) {
             bool const res = f(source);
             if (res) {
-                LOG(log_.debug()) << "Successfully executed func at source = " << source.toString()
+                LOG(log_.debug()) << "Successfully executed func at source = " << source->toString()
                                   << " - ledger sequence = " << ledgerSequence;
                 break;
             }
 
-            LOG(log_.warn()) << "Failed to execute func at source = " << source.toString()
+            LOG(log_.warn()) << "Failed to execute func at source = " << source->toString()
                              << " - ledger sequence = " << ledgerSequence;
         } else {
-            LOG(log_.warn()) << "Ledger not present at source = " << source.toString()
+            LOG(log_.warn()) << "Ledger not present at source = " << source->toString()
                              << " - ledger sequence = " << ledgerSequence;
         }
         sourceIdx = (sourceIdx + 1) % sources_.size();
@@ -307,8 +308,8 @@ LoadBalancer::chooseForwardingSource()
 {
     hasForwardingSource_ = false;
     for (auto& source : sources_) {
-        if (source.isConnected()) {
-            source.setForwarding(true);
+        if (source->isConnected()) {
+            source->setForwarding(true);
             hasForwardingSource_ = true;
             return;
         }
