+++ src/etl/Source.cpp
@@ -21,6 +21,10 @@
 
 #include "data/BackendInterface.hpp"
 #include "etl/ETLHelpers.hpp"
+#include "etl/impl/ForwardingSource.hpp"
+#include "etl/impl/GrpcSource.hpp"
+#include "etl/impl/SourceImpl.hpp"
+#include "etl/impl/SubscriptionSource.hpp"
 #include "feed/SubscriptionManagerInterface.hpp"
 #include "util/config/Config.hpp"
 
@@ -32,18 +36,16 @@
 
 namespace etl {
 
-template class SourceImpl<>;
-
-Source
+SourcePtr
 make_Source(
     util::Config const& config,
     boost::asio::io_context& ioc,
     std::shared_ptr<BackendInterface> backend,
     std::shared_ptr<feed::SubscriptionManagerInterface> subscriptions,
     std::shared_ptr<NetworkValidatedLedgersInterface> validatedLedgers,
-    Source::OnDisconnectHook onDisconnect,
-    Source::OnConnectHook onConnect,
-    Source::OnLedgerClosedHook onLedgerClosed
+    SourceBase::OnDisconnectHook onDisconnect,
+    SourceBase::OnConnectHook onConnect,
+    SourceBase::OnLedgerClosedHook onLedgerClosed
 )
 {
     auto const ip = config.valueOr<std::string>("ip", {});
@@ -63,9 +65,9 @@ make_Source(
         std::move(onLedgerClosed)
     );
 
-    return Source{
+    return std::make_unique<impl::SourceImpl<>>(
         ip, wsPort, grpcPort, std::move(grpcSource), std::move(subscriptionSource), std::move(forwardingSource)
-    };
+    );
 }
 
 }  // namespace etl
