+++ unittests/CMakeLists.txt
@@ -29,9 +29,11 @@ target_sources(
           etl/ForwardingSourceTests.cpp
           etl/GrpcSourceTests.cpp
           etl/LedgerPublisherTests.cpp
+          etl/LoadBalancerTests.cpp
           etl/SourceTests.cpp
           etl/SubscriptionSourceTests.cpp
           etl/TransformerTests.cpp
+          etl/impl/SourceImplTests.cpp
           # Feed
           feed/BookChangesFeedTests.cpp
           feed/ForwardFeedTests.cpp
